requirements.txt
Flask
PyPDF2
google-generativeai
Pillow
pdf2image
gunicorn